[DUI] Navigation: modified wording and action of enable navigation bar switch
This switch mostly duplicated the action of the similar switch in Buttons, but
was always shown whether or not a hardware navbar was available. The switch is
now only shown if a hardware navbar is available and can be disabled (just like
the one in Buttons), because it doesn't make sense to be able to disable the
software navbar otherwise.